[ahrs] mlkf: update state interface from wrapper
