+++ sw/airborne/subsystems/ahrs/ahrs_float_mlkf.h
@@ -59,7 +59,6 @@ struct AhrsMlkf {
 
   enum AhrsMlkfStatus status;
   bool_t is_aligned;
-  bool_t output_enabled; ///< if TRUE with push the estimation results to the state interface
 };
 
 extern struct AhrsMlkf ahrs_mlkf;
